Updated in line with comments
