+++ tests/tel2puml/pv_to_puml/conftest.py
@@ -1,0 +1,70 @@
+"""Fixtures for testing pv_to_puml module"""
+
+import pytest
+from datetime import datetime, timedelta
+from typing import Generator, Any
+
+from tel2puml.tel2puml_types import PVEvent
+
+
+def event_generator(
+    job_name: str,
+    job_id: str,
+    base_timestamp: datetime,
+    event_types: list[str],
+) -> Generator[PVEvent, Any, None]:
+    """Function to stream PVEvents."""
+    current_timestamp = base_timestamp
+    prev_event_id = None
+    for i in range(4):
+        event_id = f"event_{i}_{job_id}"
+        previousEventId: str | None = prev_event_id
+        yield PVEvent(
+            jobName=job_name,
+            jobId=job_id,
+            eventId=event_id,
+            eventType=event_types[i % len(event_types)],
+            applicationName="app-name",
+            timestamp=current_timestamp.isoformat(),
+            previousEventIds=previousEventId if previousEventId else [],
+        )
+        prev_event_id = event_id
+        current_timestamp += timedelta(minutes=5)
+
+
+def job_id_generator(
+    job_name: str, event_types: list[str]
+) -> Generator[Generator[PVEvent, Any, None], Any, None]:
+    """Function to stream groups of PVEvents, grouped by job_id."""
+    job_ids = [
+        f"job_1_{job_name}",
+        f"job_2_{job_name}",
+    ]
+    base_timestamp = datetime.now()
+    for job_id in job_ids:
+        yield event_generator(job_name, job_id, base_timestamp, event_types)
+        base_timestamp += timedelta(minutes=10)
+
+
+@pytest.fixture
+def pv_streams() -> Generator[
+    tuple[str, Generator[Generator[PVEvent, Any, None], Any, None]],
+    None,
+    None,
+]:
+    """Fixture to return a generator that streams 2 jobs, each with
+    2 job ids consisting of 4 PVEvents."""
+    event_types = ["A", "B", "C", "D"]
+    job_names = ["Job_A", "Job_B"]
+
+    def generator() -> Generator[
+        tuple[str, Generator[Generator[PVEvent, Any, None], Any, None]],
+        None,
+        None,
+    ]:
+        """Creates a generator of tuples of job names to a generator of
+        generators of PVEvents grouped by job name then job id."""
+        for job_name in job_names:
+            yield (job_name, job_id_generator(job_name, event_types))
+
+    return generator()
